Time: 3 ms (84.64%), Space: 16.7 MB (6.86%) - LeetHub

+++ 0152-maximum-product-subarray/0152-maximum-product-subarray.cpp
@@ -1,8 +1,8 @@
 class Solution {
 public:
     int maxProduct(vector<int>& nums) {
-        long long suffix=1;
-        long long prefix=1;
+        double suffix=1;
+        double prefix=1;
         int ret=INT_MIN;
         int n=nums.size();
         for(int i=0;i<n;i++){
@@ -10,11 +10,11 @@ public:
             if(prefix==0){prefix=1;}
             prefix*=nums[i];
             suffix*=nums[n-i-1];
-                if (prefix > INT_MAX || prefix < INT_MIN) {
-                prefix = nums[i];  // Reset prefix to current number to avoid overflow
+            if (prefix > INT_MAX || prefix < INT_MIN) {
+                prefix = nums[i];  
             }
             if (suffix > INT_MAX || suffix < INT_MIN) {
-                suffix = nums[n - i - 1];  // Reset suffix to current number to avoid overflow
+                suffix = nums[n - i - 1];
             }
             ret=max(ret,(int)max(suffix,prefix));
         }
